fix time format

+++ src/format.cpp
@@ -6,7 +6,7 @@ using std::string;
 using std::to_string;
 
 string Format::ElapsedTime(long seconds) {
-  return to_string(int(seconds / 3600)) + ':' +
-         to_string(int((seconds / 60) % 60)) + ':' +
-         to_string(int(seconds % 60));
+  return to_string(long(seconds / 3600)) + ':' +
+         to_string(long((seconds / 60) % 60)) + ':' +
+         to_string(long(seconds % 60));
 }
